1.1.0.1 Ver update (#142)
* 수집기 로그삭제 주기를 14일->7일로 변경 (#139)

* ami 장치상세팝업 안나오는 오류 수정 (#140)

* 수집기 및 서비스포털 pmgrow쪽 url 변경 및 pv 관련 api url 변경(ivtes->ivts) (#141)

pmgrow쪽 url을 최신 스펙 url로 변경
ems관련 startDt파라미터 세팅 시 timestamp를 +1 -> +1000으로 변경(0.001초->1초)

* pmgrow쪽 url 변경 및 pv 관련 api url 변경(ivtes->ivts) 오류 수정 (#143)

+++ ewp-api-agent/src/main/java/kr/co/ewp/api/entity/DeviceBms.java
@@ -15,7 +15,7 @@ public class DeviceBms {
     private Float sysSoh;//soh(단위:%)
     private Float sysVoltage;//출력전압(단위:v)
     private Float sysCurrent;//출력전류(단위:a)
-    private Integer dod;//dod(단위:%)
+    private Float dod;//dod(단위:%)
     private Date stdDate;//기준일시
     private Date regDate;//등록일시
    /**
@@ -178,14 +178,14 @@ public class DeviceBms {
     * dod(단위:%) 조회
     * @return dod
     */
-    public Integer getDod() {
+    public Float getDod() {
         return this.dod;
     }
    /**
     * dod(단위:%) 설정
     * @return dod
     */
-    public void setDod(Integer dod) {
+    public void setDod(Float dod) {
         this.dod = dod;
     }
    /**
